UI : find id form and find pw form UI setting
