+++ src/App.js
@@ -5,12 +5,13 @@ import Register from "./login/Register";
 import Auth from "./login/Auth";
 import SelectRegisterWay from "./login/SelectRegisterWay";
 import MainPage from "./map/Temp";
-import FindIdAndPw from "./login/FindIdAndPw";
 import Notice from "./Notice/NoticeBoard";
 import NoticeDetail from "./Notice/NoticeDetail";
 import NoticeCreate from "./Notice/NoticeCreate";
 import UserInfoView from "./User/UserInfoView";
 import UserInfoEdit from "./User/UserInfoEdit";
+import FindId from "./login/FindId";
+import FindPw from "./login/FindPw";
 
 function App() {
   return (
@@ -21,7 +22,8 @@ function App() {
         <Route path="/select-register" element={<SelectRegisterWay />} />
         <Route path="/default-register" element={<Register />} />
         <Route path="/kakao-register" element={<Register />} />
-        <Route path="/find-idpw" element={<FindIdAndPw />} />
+        <Route path="/find-id" element={<FindId />} />
+        <Route path="/find-pw" element={<FindPw />} />
         <Route path="/notice" element={<Notice />} />
         <Route path="/notice/view/:id" element={<NoticeDetail />} />
         <Route path="/notice/:mode/:id" element={<NoticeCreate />} />
