+++ src/style/FindIdPw.scss
@@ -1,0 +1,57 @@
+@use "util.scss" as *;
+
+.find-idpwForm {
+  font-family: "malssami815";
+  color: $main-color;
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+
+  .find-form-wrapper {
+    width: 50%;
+    height: 50%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+
+    @include size-blue-btn(100%, 15%, 2rem);
+    .find-btn {
+      margin-bottom: 30px;
+    }
+
+    .move-link {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+
+      a {
+        @include linkTextDecorate;
+        color: #666363;
+      }
+    }
+
+    input {
+      font-family: "malssami815";
+      border: 2px solid #666363;
+      padding: 10px;
+      font-size: 1.2rem;
+    }
+
+    .input-wrapper {
+      height: 27%;
+      margin: 3rem 0;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+    }
+
+    .find-input {
+      margin: 3rem 0;
+      height: 8%;
+    }
+  }
+}
