+++ src/style/util.scss
@@ -57,3 +57,18 @@ $breakpoint-large-desktop: 2048px;
     width: 140px;
   }
 }
+
+@mixin size-blue-btn($w, $h, $font) {
+  button,
+  input[type="button"] {
+    width: $w;
+    height: $h;
+    font-family: "malssami815";
+    border: none;
+    padding: 20px;
+    border-radius: 50px;
+    font-size: $font;
+    color: white;
+    background-color: $btn-blue;
+  }
+}
